+++ src/python/ap_features/__init__.py
@@ -2,12 +2,17 @@ __author__ = """Henrik Finsberg"""
 __email__ = "henriknf@simula.no"
 __version__ = "0.1.0"
 
-from . import ap_features
-from .ap_features import (
-    all_cost_terms,
+# from . import ap_features
+from ._c import (
+    NUM_COST_TERMS,
     all_cost_terms_c,
     apd_c,
     apd_up_xy_c,
+    cost_terms_c,
+    cost_terms_trace_c,
+)
+from ._numba import (
+    all_cost_terms,
     compute_APD,
     compute_APD_from_stim,
     compute_APDUpxy,
@@ -15,14 +20,11 @@ from .ap_features import (
     compute_dvdt_max,
     compute_integral,
     cost_terms,
-    cost_terms_c,
     cost_terms_trace,
-    cost_terms_trace_c,
-    list_cost_function_terms,
-    list_cost_function_terms_trace,
     peak_and_repolarization,
     transpose_trace_array,
 )
+from .utils import list_cost_function_terms, list_cost_function_terms_trace
 
 __all__ = [
     "ap_features",
@@ -44,4 +46,5 @@ __all__ = [
     "transpose_trace_array",
     "list_cost_function_terms_trace",
     "list_cost_function_terms",
+    "NUM_COST_TERMS",
 ]
